Move linking ability to its own utility (#20)
